+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEW FEATURES
 
 - Add require('@openveo/api').util.areSameArrays to shallow validates that two arrays contain the same values, no more no less
+- Add require('@openveo/api').util.evaluateDeepObjectProperties to evaluate a path of properties on an object without making use of the JavaScript eval function
 
 # 4.1.0 / 2017-09-12